PhysicsLayers bug fixed

+++ For Engine 2D/Code/Components.h
@@ -116,7 +116,7 @@ namespace FE2D {
 
         vec2 velocity = vec2();
 
-        uint32_t layer = 0;
+        uint8_t layer = 0;
 
         std::vector<Entity> entities_in;
